backend/requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.1.4
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
msgpack==1.1.2
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
firebase-admin==6.5.0 